#define SATOSHIDEN ((uint64_t)100000000L)

+++ src/cc/games/prices.c
@@ -1,7 +1,7 @@
 
 #include "prices.h"
 #include <time.h>
-#define SATOSHIDEN 100000000
+#define SATOSHIDEN ((uint64_t)100000000L)
 
 /*
  In order to port a game into gamesCC, the RNG needs to be seeded with the gametxid seed, also events needs to be broadcast using issue_games_events. Also the game engine needs to be daemonized, preferably by putting all globals into a single data structure.
